+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 2020-03-19
 
+### Added
+
+- Market access trade barriers pipeline.
+
 ### Changed
 
 - Moved datasets `fetch_from_api` to a more general `common.fetch_from_hawk_api`, as there may be a more general case for this.